+++ R/count.R
@@ -35,29 +35,30 @@ count_n <- function(df, ...) {
   dplyr::count(dplyr::count(df, ...), .data$n, name = "freq")
 }
 
-#' `dplyr::count()` with proportion column and default `sort = TRUE`.
+#' `dplyr::count()` with proportion column and default `sort = TRUE`
 #'
-#' Adds a column `pname` which gives the proportion of total rows in that group.
-#' `countp()` is an alternative name.
+#' Adds a column which gives the proportion of total rows in that group.
+#' * `vcount()` is similar but works with a vector input.
+#' * `countp()` and `countv()` are alternative names for, respectively, `count2()` and `vcount()`.
 #'
 #' @param df A data frame.
 #' @param ... Variables to group by. Passed to `count()`.
 #' @param sort Passed to `count`, but defaults to `TRUE`.
 #' @param wt,name Optional, passed to `count`.
-#' @param pname Name for the proportion column.
+#' @param pname  Name for the proportion column. If omitted, it will default to p. If there's
+#' already a column called p, it will use pp. If there's a column called p and pp, it'll use ppp,
+#' and so on, adding ps until it gets a new name.
 #'
 #' @examples
 #' count2(mtcars, cyl)
 #' count2(mtcars, cyl, sort = FALSE)
+#' vcount(c(1, 2, 2, 3))
 #'
 #' @export
-count2 <- function(df, ..., sort = TRUE, wt = NULL, name = NULL, pname = "prop") {
-  if (pname %in% names(df)){
-    stop('A column called "', pname, '" already exists in `df`. Choose a different `pname`.',
-         call. = FALSE)
-  }
+count2 <- function(df, ..., sort = TRUE, wt = NULL, name = NULL, pname = NULL) {
   x <- dplyr::count(df, ..., wt = {{wt}}, sort = sort, name = name)
   count_name <- names(x)[ncol(x)]
+  pname <- check_p_name(pname, names(x))
   dplyr::mutate(x, !!pname := .data[[count_name]] / sum(.data[[count_name]]))
 }
 
@@ -65,26 +66,37 @@ count2 <- function(df, ..., sort = TRUE, wt = NULL, name = NULL, pname = "prop")
 #' @export
 countp <- count2
 
-#' Frequency count for values of a vector with tibble output.
-#'
-#' This behaves similarly to `count2()` but with a vector input. `countv()` is an alternative name.
-#'
-#' @param x An atomic vector.
-#' @param sort Logical. Sort output by descending frequency. If `FALSE` sort by value.
-#' @param name Name of count column in output. If omitted `n` will be used (as in `dplyr::count`).
-#' @param value_name Name of value column.
-#'
+#' @param x A vector (including a data frame).
+#' @seealso [vctrs::vec_count()]
+#' @rdname count2
 #' @export
-vcount <- function(x, sort = TRUE, name = NULL, value_name = "value") {
-  if (!is.null(name) && (value_name == name)){
-    stop("`name` and `value_name` must not be the same.", call. = FALSE)
-  }
-  res <- tibble::as_tibble_col(x, value_name) %>%
-    dplyr::count(.data[[value_name]], sort = sort, name = name)
-  count_name <- names(res)[ncol(res)]
-  dplyr::mutate(res, prop = .data[[count_name]] / sum(.data[[count_name]]))
+vcount <- function(x, sort = TRUE, name = NULL, pname = NULL) {
+  dat <- tibble::as_tibble_col(x, "value")
+  count2(dat, .data$value, sort = sort, name = name, pname = pname)
 }
 
-#' @rdname vcount
+#' @rdname count2
 #' @export
 countv <- vcount
+
+# helpers ---------
+# Similar to functions check_n_name() and n_name() in dplyr. See dplyr/R/count-tally.R
+check_p_name <- function(name, vars) {
+  if (is.null(name)) {
+    name <- p_name(vars)
+
+    if (name != "p") {
+      rlang::inform(c(paste0("Storing counts in `", name, "`, as `p` already present in input"),
+        i = "Use `name = \"new_name\"` to pick a new name."))
+    }
+  }
+  name
+}
+
+p_name <- function(x) {
+  name <- "p"
+  while (name %in% x) {
+    name <- paste0("p", name)
+  }
+  name
+}